requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic-settings==2.10.1
uvicorn==0.35.0
sqlalchemy==2.0.49
alembic==1.18.4
psycopg[binary]==3.2.13
asyncpg==0.30.0
greenlet==3.2.4
redis==5.0.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
boto3>=1.34.0,<2.0.0
pytest==7.4.3
pytest-asyncio==0.21.1